requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.19.0
psycopg2-binary==2.9.1
resend==1.0.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
